requirements.txt
numpy~=1.21.2
opencv-contrib-python~=4.5.3
scipy~=1.7.1
matplotlib~=3.4.3
pygame~=2.0.2
tensorflow~=2.7.0
Pillow~=8.4.0